ReadME
Automates CTR for Cardiomegaly using Deep Learning Approach.
This project automates the detection of cardiomegaly (enlarged heart) using a deep learning approach. The project consists of a Python script that trains a convolutional neural network (CNN) on a dataset of chest X-rays. The trained CNN can then be used to classify new chest X-rays as having or not having cardiomegaly. The cardiothoracic ratio (CTR), expressing the relationship between the size of the heart and the transverse dimension of the chest measured on a chest PA radiograph, is a commonly used parameter in the assessment of cardiomegaly with a cut-off value of 0.5.
Features:
1. Automates the calculation of CTR and detection of cardiomegaly in chest X-rays.
2. Can save time and improve the accuracy of CTR calculation and cardiomegaly detection.
3. Uses a deep learning model that is trained on a large dataset of chest X-rays.
Requirements:
Python 3, Numpy, Pandas, Matplotlib, Tensorflow, Keras.